refactor(ui): address scheduled tasks review feedback
